+++ logging/hooks/custom_logs.go
@@ -3,10 +3,12 @@ package hooks
 import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/sirupsen/logrus"
+	"github.com/zclconf/go-cty/cty"
 
 	"github.com/avenga/couper/config/env"
 	"github.com/avenga/couper/config/request"
 	"github.com/avenga/couper/eval"
+	"github.com/avenga/couper/internal/seetie"
 	"github.com/avenga/couper/logging"
 )
 
@@ -76,72 +78,18 @@ func fireAccess(entry *logrus.Entry) {
 		return
 	}
 
-	ctx := syncedUpstreamContext(evalCtx, entry)
-
-	if fields := eval.ApplyCustomLogs(ctx, hclBodies, entry); len(fields) > 0 {
+	if fields := eval.ApplyCustomLogs(evalCtx.HCLContextSync(), hclBodies, entry); len(fields) > 0 {
 		entry.Data[customLogField] = fields
 	}
 }
 
 func fireUpstream(entry *logrus.Entry) {
-	evalCtx, ok := entry.Context.Value(request.ContextType).(*eval.Context)
-	if !ok {
+	logError, _ := entry.Context.Value(request.LogCustomUpstreamError).(*error)
+	if *logError != nil {
+		entry.Debug(*logError)
 		return
 	}
-	bodyCh, _ := entry.Context.Value(request.LogCustomUpstream).(chan hcl.Body)
-	if bodyCh == nil {
-		return
-	}
-
-	var bodies []hcl.Body
-	select {
-	case body := <-bodyCh:
-		bodies = append(bodies, body)
-	case <-entry.Context.Done():
-	}
-
-	ctx := syncedUpstreamContext(evalCtx, entry)
-
-	if fields := eval.ApplyCustomLogs(ctx, bodies, entry); len(fields) > 0 {
-		entry.Data[customLogField] = fields
-	}
-}
-
-// syncedUpstreamContext prepares the local backend variable.
-func syncedUpstreamContext(evalCtx *eval.Context, entry *logrus.Entry) *hcl.EvalContext {
-	ctx := evalCtx.HCLContextSync()
-
-	tr, _ := entry.Context.Value(request.TokenRequest).(string)
-	rtName, _ := entry.Context.Value(request.RoundTripName).(string)
-	isTr := tr != ""
-
-	if rtName == "" {
-		return ctx
-	}
-
-	if _, ok := ctx.Variables[eval.BackendRequests]; ok {
-		for k, v := range ctx.Variables[eval.BackendRequests].AsValueMap() {
-			if isTr && k == eval.TokenRequestPrefix+tr {
-				ctx.Variables[eval.BackendRequest] = v
-				break
-			} else if k == rtName {
-				ctx.Variables[eval.BackendRequest] = v
-				break
-			}
-		}
-	}
-
-	if _, ok := ctx.Variables[eval.BackendResponses]; ok {
-		for k, v := range ctx.Variables[eval.BackendResponses].AsValueMap() {
-			if isTr && k == eval.TokenRequestPrefix+tr {
-				ctx.Variables[eval.BackendResponse] = v
-				break
-			} else if k == rtName {
-				ctx.Variables[eval.BackendResponse] = v
-				break
-			}
-		}
-	}
-
-	return ctx
+	logValue, _ := entry.Context.Value(request.LogCustomUpstreamValue).(*cty.Value)
+	fields := seetie.ValueToLogFields(*logValue)
+	entry.Data[customLogField] = fields
 }