+++ logging/upstream_log.go
@@ -20,7 +20,6 @@ import (
 	"github.com/avenga/couper/handler/validation"
 	"github.com/avenga/couper/internal/seetie"
 	"github.com/avenga/couper/utils"
-	"github.com/hashicorp/hcl/v2"
 )
 
 var (
@@ -82,10 +81,12 @@ func (u *UpstreamLog) RoundTrip(req *http.Request) (*http.Response, error) {
 
 	fields["request"] = requestFields
 
+	var logValue cty.Value
+	var logError error
 	berespBytes := int64(0)
-	logCtxCh := make(chan hcl.Body, 17) // TODO: Will block with oauth2 token retries >= 17
 	tokenRetries := uint8(0)
-	outctx := context.WithValue(req.Context(), request.LogCustomUpstream, logCtxCh)
+	outctx := context.WithValue(req.Context(), request.LogCustomUpstreamValue, &logValue)
+	outctx = context.WithValue(outctx, request.LogCustomUpstreamError, &logError)
 	outctx = context.WithValue(outctx, request.BackendBytes, &berespBytes)
 	outctx = context.WithValue(outctx, request.TokenRequestRetries, &tokenRetries)
 	oCtx, openAPIContext := validation.NewWithContext(outctx)
@@ -94,8 +95,6 @@ func (u *UpstreamLog) RoundTrip(req *http.Request) (*http.Response, error) {
 	rtStart := time.Now()
 	beresp, err := u.next.RoundTrip(outreq)
 	rtDone := time.Now()
-
-	close(logCtxCh)
 
 	// FIXME: Can host be empty?
 	if outreq.Host != "" {